--- apps/platform/src/app/page.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-export default function Home() {
-
-  return (
-    <main>
-      <h1>Welcome to platform</h1>
-    </main>
-  );
-}
