+++ apps/supabase-api/Makefile
@@ -64,16 +64,18 @@ setup: clean-supabase
 
 clean-supabase: ## Stop Supabase and clean volumes if running, started, or in an unhealthy state
 	@echo "Checking Supabase status..."
-	@supabase status
+	@supabase status || true
 	@echo "Supabase status check complete."
 	@echo "Evaluating condition..."
-	@if supabase status | grep -q . ; then \
+	@if supabase status 2>/dev/null | grep -q . ; then \
 		echo "Condition is TRUE"; \
 		echo "Detected Supabase running or in problematic state."; \
 		echo "Stopping Supabase and cleaning volumes..."; \
 		supabase stop || true; \
 		echo "Stopping any remaining Supabase containers..."; \
-		docker volume rm $$(docker volume ls -q --filter label=com.supabase.cli.project=trnlccdjeihaaxmcaddm) || true; \
+		docker stop supabase_db_trnlccdjeihaaxmcaddm >/dev/null 2>&1 || true; \
+		docker rm supabase_db_trnlccdjeihaaxmcaddm >/dev/null 2>&1 || true; \
+		docker volume rm $$(docker volume ls -q --filter label=com.supabase.cli.project=trnlccdjeihaaxmcaddm) >/dev/null 2>&1 || true; \
 		echo "Supabase stopped and volumes cleaned."; \
 	else \
 		echo "Condition is FALSE"; \
