--- turbo/generator/templates/app/src/app/loading.tsx.hbs
@@ -1,3 +1,0 @@
-export default function Loading() {
-  return <div>Loading...</div>;
-}
