--- turbo/generator/templates/app/src/app/page.tsx.hbs
@@ -1,10 +1,0 @@
-import React from "react";
-
-export default function Home() {
-
-  return (
-    <main>
-      <h1>Welcome to {{kebabCase name}}</h1>
-    </main>
-  );
-}
